Fix hyperlink post summary issue (#30)
According to the HTML spec, <a> tags cannot be nested. Therefore, when
the content of the post includes an <a> tag, it causes strange browser
interactions.

To fix this, we simply remove all html tags from the post excerpt before
displaying it, so that there will never be links inside the post.