Temporary Christmas card page (#88)
* Temporary Christmas card page

All in one directory, to be used temporarily as a replacement for the cssbham.com/christmas/ url. It contains all the necessary files (images, HTML, CSS, JS) including a noscript version of the main thing which index.html redirects to if it identifies that JS is disabled.

* Added favicons

* Indents added

Easier readability

* Changed styling

From inline to file.